requirements.txt
pyedifice==0.3.0
numpy==1.26.2
pydub==0.25.1
simpleaudio==1.0.4
pyaudio
jsonpickle
pythonlangutil
typing_extensions
typing-extensions
PySide6
tinytag
pynput